backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
google-generativeai==0.3.1
openai==1.3.0
python-multipart==0.0.6
httpx==0.25.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1 